% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_all.R
\name{DD_dataCarlinLouis}
\alias{DD_dataCarlinLouis}
\title{this will download files from https://www.counterpointstat.com/bayesian-methods-for-data-analysis.html and install the package dataCarlinLouis}
\usage{
DD_dataCarlinLouis()
}
\description{
this will download files from https://www.counterpointstat.com/bayesian-methods-for-data-analysis.html and install the package dataCarlinLouis
}
